move shared sr functions to pkg
